+++ pkg/cmd/login/login.go
@@ -1,5 +1,12 @@
 package login
 
+import (
+	"fmt"
+	"os"
+	"strconv"
+)
+
+// NewLogin login to the selected instance.
 func NewLogin(env string, app string, name string, region string, user string, silent bool, ssh bool) {
 	rawInstanceList := filterInstances(region, env, app, name, silent, ssh)
 	instances := getInstancesInfo(rawInstanceList)
@@ -17,4 +24,42 @@ func NewLogin(env string, app string, name string, region string, user string, s
 	} else {
 		launchSSM(instances[selectedInstance].ID)
 	}
+}
+
+func showIPsList(instanceList []Instance) {
+	if len(instanceList) == 0 {
+		os.Exit(0)
+	}
+	for _, inst := range instanceList {
+		fmt.Printf("%s\n", inst.IP)
+	}
+	os.Exit(0)
+}
+
+func showInstanceList(instanceList []Instance, user string) {
+	if len(instanceList) == 0 {
+		fmt.Printf("There are no instances matching your request.\n")
+		os.Exit(0)
+	}
+	for idx, inst := range instanceList {
+		fmt.Printf("[%d] %s - %s (%s, %s) \n", idx, inst.Name, inst.IP, inst.ID, inst.Size)
+	}
+}
+
+func selectInstanceIndex(instanceList []Instance) int {
+	var input string
+	var index int
+	if len(instanceList) == 1 {
+		fmt.Printf("\n\n")
+		index = 0
+	} else {
+		fmt.Println("\nWhich one do you want to ssh in?")
+		fmt.Scanln(&input)
+		index, err := strconv.Atoi(input)
+		if err != nil || index > len(instanceList)-1 || index < 0 {
+			fmt.Println("Please enter a valid number.")
+			index = selectInstanceIndex(instanceList)
+		}
+	}
+	return index
 }
