+++ pkg/cmd/version.go
@@ -1,7 +1,6 @@
 package cmd
 
+// Get Returns the version of the library.
 func Get() string {
-	var version string
-	version = "v1.2.7"
-	return version
+	return "v1.2.7"
 }
